Various improvements/fixes
