+++ .env.test
@@ -3,3 +3,4 @@ CLIENT_ID=111398433321891634
 BOT_TOKEN=AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
 TEST=true
 CI=true
+YARN_ENABLE_HARDENED_MODE=0
