+++ .github/workflows/unit_tests.yml
@@ -11,7 +11,7 @@ jobs:
   test:
     name: Node v${{ matrix.node_version }} - ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
-    timeout-minutes: 10
+    timeout-minutes: 5
     strategy:
       matrix:
         node_version: [20.15.0]
