+++ src/mahoji/lib/inhibitors.ts
@@ -1,8 +1,7 @@
+import { PerkTier, formatDuration } from '@oldschoolgg/toolkit';
 import type { DMChannel, Guild, GuildMember, InteractionReplyOptions, TextChannel, User } from 'discord.js';
 import { ComponentType, PermissionsBitField } from 'discord.js';
 
-import { PerkTier } from '@oldschoolgg/toolkit';
-import { formatDuration } from '@oldschoolgg/toolkit';
 import { OWNER_IDS, SupportServer } from '../../config';
 import { BLACKLISTED_GUILDS, BLACKLISTED_USERS } from '../../lib/blacklists';
 import { BadgesEnum, BitField, Channel, DISABLED_COMMANDS, minionBuyButton } from '../../lib/constants';
